+++ backend/company_app/views.py
@@ -2,8 +2,11 @@ from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import status
 from django.shortcuts import get_object_or_404
+
+from client_app.models import Client
 from .models import Company
 from .serializers import CompanySerializer
+
 class CompanyView(APIView):
     
     def get(self, request, pk=None):
@@ -38,8 +41,18 @@ class CompanyView(APIView):
         company = get_object_or_404(Company, id=pk)
         hard = request.query_params.get("hard", "false").lower() == "true"
         if hard:
+            if request.user.role != "admin":
+                raise self.permission_denied(request,
+                message="Solo admin puede realizar borrado físico."
+            );
             company.delete()
             return Response(status=status.HTTP_204_NO_CONTENT)    
+        
+        relatedClients = Client.objects.filter(company=company, enabled=True);
+        
+        if (relatedClients.exists()):
+            return Response({"message": "No se puede eliminar una empresa que tiene clientes asociados. Elimine o reasigne los clientes primero."}, status=status.HTTP_400_BAD_REQUEST);
+        
         company.enabled = False
         company.save()
         return Response(status=status.HTTP_204_NO_CONTENT)
